requirements.txt
asgiref==3.4.1
boto3==1.18.44
botocore==1.21.44
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
Deprecated==1.2.13
Django==3.2.7
django-filter==2.4.0
django-storages==1.11.1
gcloud==0.18.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
jmespath==0.10.0
jwcrypto==1.0
oauth2client==4.1.3
Pillow==8.3.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-jwt==3.3.0
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.6
whitenoise==5.3.0
wrapt==1.12.1
